fix(SolrSearch): Changed default search type from 'Page' to 'SiteTree' to fix a bug with Mediawesome, as its MediaPage class extends SiteTree, rather than Page. Made the default search type configurable as well.
NOTE: There are a couple of calls to 'searchableTypes' that pass in 'Page' and may need to be factored in the future to work with this config. I only changed the default 'Page' value for functions that were within the 'getQuery' code path as that is all that's been tested.